+++ src/Bootstrap4ProgressBarMultiple.php
@@ -3,8 +3,7 @@ declare(strict_types=1);
 
 namespace DBlackborough\Zf3ViewHelpers;
 
-use PHPUnit\Runner\Exception;
-use Zend\View\Helper\AbstractHelper;
+use Exception;
 
 /**
  * Generate a Bootstrap 4 progress bar
@@ -14,17 +13,12 @@ use Zend\View\Helper\AbstractHelper;
  * @copyright Dean Blackborough
  * @license https://github.com/deanblackborough/zf3-view-helpers/blob/master/LICENSE
  */
-class Bootstrap4ProgressBarMultiple extends AbstractHelper
+class Bootstrap4ProgressBarMultiple extends Bootstrap4Helper
 {
     /**
      * @var array Progress bar values
      */
     private $values;
-
-    /**
-     * @var array Background colours for progress bars
-     */
-    private $colors;
 
     /**
      * @var boolean Enabled the striped style
@@ -37,73 +31,30 @@ class Bootstrap4ProgressBarMultiple extends AbstractHelper
     private $animate;
 
     /**
-     * @var string Label for progress bar
-     */
-    private $label;
-
-    /**
      * @var integer Height of progress bar in pixels
      */
     private $height;
 
     /**
-     * @var array Bootstrap styles
+     * @var array Background colors
      */
-    private $supported_styles = [
-        'success',
-        'info',
-        'warning',
-        'danger',
-    ];
+    private $bg_colors;
 
     /**
      * Entry point for the view helper
      *
      * @param array $values Progress bar values, array of integers
+     * @param array $colors Utility colours to use for each progress bar
      *
      * @return Bootstrap4ProgressBarMultiple
      */
-    public function __invoke(array $values): Bootstrap4ProgressBarMultiple
+    public function __invoke(array $values, array $colors): Bootstrap4ProgressBarMultiple
     {
         $this->reset();
 
         $this->values = $values;
 
-        return $this;
-    }
-
-    /**
-     * Set the background colors for the progress bar, each value in the array needs to be one of the
-     * following, success, info, warning or danger. If an incorrect style is passed in we don't apply the
-     * class. The number of entries in the array needs to match the number of progress bar values
-     *
-     * @param array $colors
-     *
-     * @return Bootstrap4ProgressBarMultiple
-     */
-    public function color(array $colors): Bootstrap4ProgressBarMultiple
-    {
-        if (count($colors) !== count($this->values)) {
-            throw new Exception('Number of progress bar colours needs to match the number or progress bar values');
-        }
-
-        foreach ($colors as $color) {
-            if (in_array($color, $this->supported_styles) === true) {
-                $this->colors[] = $color;
-            }
-        }
-
-        return $this;
-    }
-
-    /**
-     * Enable the striped style for the progress bar background
-     *
-     * @return Bootstrap4ProgressBarMultiple
-     */
-    public function striped() : Bootstrap4ProgressBarMultiple
-    {
-        $this->striped = true;
+        $this->bgStyles($colors);
 
         return $this;
     }
@@ -121,17 +72,24 @@ class Bootstrap4ProgressBarMultiple extends AbstractHelper
     }
 
     /**
-     * Set the label for the progress bar
+     * Set the background colour for each of the progress bars, values need to be one of the following, primary,
+     * secondary, success, danger, warning, info, light, dark or white
      *
-     * @param string $label
+     * @param array $colors
      *
-     * @return Bootstrap4ProgressBarMultiple
+     * @throws Exception
      */
-    public function label(string $label) : Bootstrap4ProgressBarMultiple
+    private function bgStyles(array $colors) : void
     {
-        $this->label = $label;
+        if (count($colors) !== count($this->values)) {
+            throw new Exception('Number of progress bar colours needs to match the number or progress bar values');
+        }
 
-        return $this;
+        foreach ($colors as $color) {
+            if (in_array($color, $this->supported_bg_styles) === true) {
+                $this->bg_colors[] = $color;
+            }
+        }
     }
 
     /**
@@ -141,11 +99,74 @@ class Bootstrap4ProgressBarMultiple extends AbstractHelper
      *
      * @return Bootstrap4ProgressBarMultiple
      */
-    public function height(int $height) : Bootstrap4ProgressBarMultiple
+    public function setHeight(int $height) : Bootstrap4ProgressBarMultiple
     {
         $this->height = $height;
 
         return $this;
+    }
+
+    /**
+     * Enable the striped style for the progress bar background
+     *
+     * @return Bootstrap4ProgressBarMultiple
+     */
+    public function striped() : Bootstrap4ProgressBarMultiple
+    {
+        $this->striped = true;
+
+        return $this;
+    }
+
+    /**
+     * Generate the additional classes
+     *
+     * @param string $color
+     *
+     * @return string
+     */
+    private function classes(string $color) : string
+    {
+        $classes = '';
+
+        if ($this->striped === true) {
+            $classes .= ' progress-bar-striped';
+        }
+
+        if ($this->animate === true) {
+            $classes .= ' progress-bar-animated';
+        }
+
+        if ($color !== null) {
+            $classes .= ' bg-' . $color;
+        }
+
+        return $classes;
+    }
+
+    /**
+     * Worker method for the view helper, generates the HTML, the method is private so that we
+     * can echo/print the view helper directly
+     *
+     * @return string
+     */
+    protected function render() : string
+    {
+        $html = '<div class="progress">';
+
+        foreach ($this->values as $k => $value) {
+            $styles = $this->styles($value);
+            if (strlen($styles) > 0) {
+                $styles = 'style="' . trim($styles) . '"';
+            }
+
+            $html .= '<div class="progress-bar' . $this->classes($this->bg_colors[$k]) . '" role="progressbar" ' .
+                $styles . ' aria-valuenow="' . $value . '" aria-valuemin="0" aria-valuemax="100"></div>';
+        }
+
+        $html .= '</div>';
+
+        return $html;
     }
 
     /**
@@ -156,7 +177,7 @@ class Bootstrap4ProgressBarMultiple extends AbstractHelper
     private function reset(): void
     {
         $this->values = [];
-        $this->colors = [];
+        $this->bg_colors = [];
         $this->striped = false;
         $this->animate = false;
         $this->label = null;
@@ -179,72 +200,9 @@ class Bootstrap4ProgressBarMultiple extends AbstractHelper
         }
 
         if ($this->height !== null && $this->height > 0) {
-            $styles .= ' height: ' . $this->height . 'px;';
+            $styles .= ' height:' . $this->height . 'px;';
         }
 
         return $styles;
-    }
-
-    /**
-     * Generate the additional classes
-     *
-     * @param string $color
-     *
-     * @return string
-     */
-    private function classes(string $color) : string
-    {
-        $classes = '';
-
-        if ($color !== null) {
-            $classes .= ' bg-' . $color;
-        }
-
-        if ($this->striped === true) {
-            $classes .= ' progress-bar-striped';
-        }
-
-        if ($this->animate === true) {
-            $classes .= ' progress-bar-animated';
-        }
-
-        return $this->view->escapeHtmlAttr($classes);
-    }
-
-    /**
-     * Worker method for the view helper, generates the HTML, the method is private so that we
-     * can echo/print the view helper directly
-     *
-     * @return string
-     */
-    private function render() : string
-    {
-        $html = '<div class="progress">';
-
-        foreach ($this->values as $k => $value) {
-            $styles = $this->styles($value);
-            if (strlen($styles) > 0) {
-                $styles = 'style="' . $this->view->escapeHtmlAttr(trim($styles)) . '"';
-            }
-
-            $html .= '<div class="progress-bar' . $this->classes($this->colors[$k]) . '" role="progressbar" ' .
-                $styles . ' aria-valuenow="' . $value . '" aria-valuemin="0" aria-valuemax="100">' .
-                (($this->label !== null) ? $this->view->escapeHtml($this->label) : null) . '</div>' . PHP_EOL;
-        }
-
-        $html .= '</div>';
-
-        return $html;
-    }
-
-    /**
-     * Override the __toString() method to allow echo/print of the view helper directly,
-     * saves a call to render()
-     *
-     * @return string
-     */
-    public function __toString() : string
-    {
-        return $this->render();
     }
 }