+++ tests/Bootstrap4BadgeTest.php
@@ -1,7 +1,6 @@
 <?php
 
 require __DIR__ . '../../vendor/autoload.php';
-require __DIR__ . '../../src/Bootstrap4Badge.php';
 
 use DBlackborough\Zf3ViewHelpers\Bootstrap4Badge;
 
@@ -14,7 +13,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     {
         $view_helper = new Bootstrap4Badge();
         $view_helper->__invoke('Badge');
-        $this->assertEquals('<span class="badge badge-primary">Badge</span>', $view_helper->__toString());
+        $this->assertEquals('<span class="badge">Badge</span>', $view_helper->__toString());
     }
 
     /**
@@ -24,7 +23,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     {
         $view_helper = new Bootstrap4Badge();
         $view_helper->__invoke('Badge')->pill();
-        $this->assertEquals('<span class="badge badge-primary badge-pill">Badge</span>', $view_helper->__toString());
+        $this->assertEquals('<span class="badge badge-pill">Badge</span>', $view_helper->__toString());
     }
 
     /**
@@ -33,7 +32,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStylePrimary()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('primary');
+        $view_helper->__invoke('Badge')->setBgStyle('primary');
         $this->assertEquals('<span class="badge badge-primary">Badge</span>', $view_helper->__toString());
     }
 
@@ -43,7 +42,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleSecondary()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('secondary');
+        $view_helper->__invoke('Badge')->setBgStyle('secondary');
         $this->assertEquals('<span class="badge badge-secondary">Badge</span>', $view_helper->__toString());
     }
 
@@ -53,7 +52,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleSuccess()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('success');
+        $view_helper->__invoke('Badge')->setBgStyle('success');
         $this->assertEquals('<span class="badge badge-success">Badge</span>', $view_helper->__toString());
     }
 
@@ -63,7 +62,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleDanger()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('danger');
+        $view_helper->__invoke('Badge')->setBgStyle('danger');
         $this->assertEquals('<span class="badge badge-danger">Badge</span>', $view_helper->__toString());
     }
 
@@ -73,7 +72,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleWarning()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('warning');
+        $view_helper->__invoke('Badge')->setBgStyle('warning');
         $this->assertEquals('<span class="badge badge-warning">Badge</span>', $view_helper->__toString());
     }
 
@@ -83,7 +82,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleWInfo()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('info');
+        $view_helper->__invoke('Badge')->setBgStyle('info');
         $this->assertEquals('<span class="badge badge-info">Badge</span>', $view_helper->__toString());
     }
 
@@ -93,7 +92,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleLight()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('light');
+        $view_helper->__invoke('Badge')->setBgStyle('light');
         $this->assertEquals('<span class="badge badge-light">Badge</span>', $view_helper->__toString());
     }
 
@@ -103,7 +102,7 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleDark()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('dark');
+        $view_helper->__invoke('Badge')->setBgStyle('dark');
         $this->assertEquals('<span class="badge badge-dark">Badge</span>', $view_helper->__toString());
     }
 
@@ -113,8 +112,8 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     public function testStyleInvalidOption()
     {
         $view_helper = new Bootstrap4Badge();
-        $view_helper->__invoke('Badge')->setStyle('invalid');
-        $this->assertEquals('<span class="badge badge-primary">Badge</span>', $view_helper->__toString());
+        $view_helper->__invoke('Badge')->setBgStyle('invalid');
+        $this->assertEquals('<span class="badge">Badge</span>', $view_helper->__toString());
     }
 
     /**
@@ -124,6 +123,24 @@ final class Bootstrap4BadgeTest extends \PHPUnit\Framework\TestCase
     {
         $view_helper = new Bootstrap4Badge();
         $view_helper->__invoke('Badge')->asLink('#');
-        $this->assertEquals('<a href="#" class="badge badge-primary">Badge</a>', $view_helper->__toString());
+        $this->assertEquals('<a href="#" class="badge">Badge</a>', $view_helper->__toString());
     }
+    
+    /**
+     * Combined test
+     */
+    public function testCombined()
+    {
+        $view_helper = new Bootstrap4Badge();
+        $view_helper->__invoke('Badge')
+            ->asLink('#')
+            ->pill()
+            ->setBgStyle('primary')
+            ->setTextStyle('light');
+        $this->assertEquals(
+            '<a href="#" class="badge badge-primary text-light badge-pill">Badge</a>',
+            $view_helper->__toString()
+        );
+    }
+    
 }