+++ tests/Bootstrap4ProgressBar.php
@@ -1,7 +1,6 @@
 <?php
 
 require __DIR__ . '../../vendor/autoload.php';
-require __DIR__ . '../../src/Bootstrap4Jumbotron.php';
 
 use DBlackborough\Zf3ViewHelpers\Bootstrap4ProgressBar;
 
@@ -66,7 +65,7 @@ final class Bootstrap4ProgressBarTest extends PHPUnit\Framework\TestCase
     {
         $view_helper = new Bootstrap4ProgressBar();
         $view_helper->__invoke(25)
-            ->label('progress');
+            ->setLabel('progress');
         $this->assertEquals(
             '<div class="progress"><div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100">progress</div></div>',
             $view_helper->__toString()
@@ -80,7 +79,7 @@ final class Bootstrap4ProgressBarTest extends PHPUnit\Framework\TestCase
     {
         $view_helper = new Bootstrap4ProgressBar();
         $view_helper->__invoke(25)
-            ->height(5);
+            ->setHeight(5);
         $this->assertEquals(
             '<div class="progress"><div class="progress-bar" role="progressbar" style="width: 25%; height: 5px;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100"></div></div>',
             $view_helper->__toString()
@@ -378,6 +377,24 @@ final class Bootstrap4ProgressBarTest extends PHPUnit\Framework\TestCase
             ->setBgStyle('invalid');
         $this->assertEquals(
             '<div class="progress"><div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100"></div></div>',
+            $view_helper->__toString()
+        );
+    }
+
+    /**
+     * Combined test
+     */
+    public function testCombined()
+    {
+        $view_helper = new Bootstrap4ProgressBar();
+        $view_helper->__invoke(25)
+            ->setBgStyle('primary')
+            ->setTextStyle('light')
+            ->setLabel('label')
+            ->striped()
+            ->animate();
+        $this->assertEquals(
+            '<div class="progress"><div class="progress-bar bg-primary text-light progress-bar-striped progress-bar-animated" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100">label</div></div>',
             $view_helper->__toString()
         );
     }